requirements.txt
PyQt6
pandas
matplotlib
openpyxl
Pillow
reportlab
reportlab

